Change; Move all logic into event
